requirements.txt
Flask==2.3.3
google-auth==2.18.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.113.0
gunicorn==20.1.0
flask-restx==1.1.0